% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.r
\name{print.robxp}
\alias{print.robxp}
\title{Custom printer for 'robexp' objects}
\usage{
\method{print}{robxp}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Custom printer for 'robexp' objects
}
