% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sitemaps}
\alias{sitemaps}
\title{Retrieve a character vector of sitemaps from a parsed robots.txt object}
\usage{
sitemaps(xp)
}
\arguments{
\item{xp}{A \code{robxp} object}
}
\value{
charcter vector of all sitemaps found in the parsed \code{robots.txt} file
}
\description{
Retrieve a character vector of sitemaps from a parsed robots.txt object
}
\examples{
imdb <- paste0(readLines(system.file("extdata", "imdb-robots.txt",
               package="rep")), collapse="\n")
rt <- robxp(imdb)
sitemaps(rt)
}
