% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robxp.r
\name{print.robxp}
\alias{print.robxp}
\title{Custom printer for `robxp`` objects}
\usage{
\method{print}{robxp}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Custom printer for `robxp`` objects
}
\keyword{internal}
