% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.r
\name{robxp}
\alias{robxp}
\title{Create a robots.txt object}
\usage{
robxp(x)
}
\arguments{
\item{x}{atomic character vector containing a complete robots.txt file}
}
\description{
Create a robots.txt object
}
\examples{
library(robotstxt)
can_fetch(rt, "/asthma/asthma_stats/default.htm", "*") # TRUE
can_fetch(rt, "/_borders", "*") # FALSE
}
